Organizar código.
He fixeado el código del archivo productos_tab, factorizándolo, organizándolo, dejandolo todo más limpio y organizado.
