+++ lib/screens/login/login_screen.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_application_1/core/app_colors.dart';
 import 'package:firebase_auth/firebase_auth.dart'; // Necesario para login real
-import 'package:flutter_application_1/core/firebase/auth_service.dart';
+import 'package:flutter_application_1/services/firebase/auth_service.dart';
 
 class LoginScreen extends StatefulWidget {
   const LoginScreen({super.key});
