+++ lib/screens/user_dashboards/user_dashboard.dart
@@ -1,9 +1,11 @@
 
 
+// ignore_for_file: library_private_types_in_public_api
+
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_application_1/core/app_colors.dart';
-import 'package:flutter_application_1/screens/login/carrito_screen.dart';
+import 'package:flutter_application_1/screens/user_dashboards/carrito_screen.dart';
 
 class UserDashboard extends StatefulWidget {
   const UserDashboard({super.key});
